feat: reviewed-totals-layout
