+++ frontend/web/styles/project/_layout.scss
@@ -25,3 +25,7 @@
 .gap-x-3 {
   column-gap: 3rem;
 }
+
+.gap-x-2 {
+  column-gap: 0.5rem;
+}
